+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Apply `Shantell Sans` font family for `Wireframe` theme instead of `Chalkboard SE` (Orange-OpenSource/ouds-ios#990)
 - Icon for components catalog documentation (Orange-OpenSource/ouds-ios#1019)
 - `Color` raw tokens (tokens library v1.6.0) (Orange-OpenSource/ouds-ios#991)
 - `Font` semantic tokens (tokens library v1.6.0) (Orange-OpenSource/ouds-ios#991)